Begin to style intro section
As of the current state there are some placement issues with the header
and content class. Both elements aren't aligned with each other
vertically. Their child elements also need to be worked on with
alignment. The placeholder image was added to the info section however
still lacks a valid width or height.